backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
